Small refactorings of Pass/mapChildren
(1a) PostPass was an attempt at the "recursion strategy" pattern, and had some useful tricks, like it automated away some of the details of results/location, let you throw results, etc.  But not every pass that could use that automation was actually a post-order traversal, so I made "CatchingPass" that was similar, but not a post-order traversal.

We rarely used CatchingPass, though, so I just put that functionality into Pass itself.

(1b) Then, because PostPass now duplicated some of the functionality of its ancestor, I had removed some of that from PostPass and had its ancestor take care of it instead.

(1c) PostPass also allowed for the construction of pre-order traversals, and circum-order traversals, so the name was a bit misleading.  It's now "AutoPass".

(1d) I always kept forgetting where special passes were -- they weren't in `passes.ts`, they were in `passes/ancestorPasses.ts`.  There was some reason for this but I can't remember why.  (I suspect there was a JavaScript recursive import problem, those are difficult to fix without separating things up sometimes, but whatever it was it's not an issue now.)  Anyway, the latter file is gone now, everything's in `passes.ts` now.

(2) We used to treat members starting with "_" as special and ephemeral.  `mapChildren` doesn't clone them, `toStr` doesn't print them, etc.  But one by one we've gotten rid of all ephemeral members, so this wasn't doing anything anymore, and could have been a footgun in the future, so I removed it.