Calculating tapes and vocabs together
I think it's best if we calculate tapes and vocabs together in one fell swoop, otherwise we're solving the same tricky question ("What does this embed refer to and do I know everything I need to know about it yet?") multple times.

For now I've only worked out a limited slice of the system.

(1) Only string-type vocabs, I'm turning off the atomicity optimization for now.

(2) There are a bunch of cases (e.g., SingleTape, Match, ReplaceBlock) in the current tape calculation where we know the NAME of a tape early, but that's all we know.  We'll need additional types of suspension to handle this.