Fixing strange bug with JS Maps
We got a very strange bug where `TapeLit.tapes: Map<string,VocabString>` was no sometimes longer a compatible receiver for Map.prototype.keys() (or any of the Map methods).  I checked and its type and constructor are still `Map`, so if `.keys()` is unavailable on it, it suggests that the prototype has changed.  But we never mess around with prototypes.  Could the test framework be mocking Map somehow?

Anyway, I never figured it out, but changing these back to the JS Objects that we're used to (that is, `Dict<VocabString>`) got rid of this problem.