Getting rid of Intersection
Intersection actually works differently than Join for the purposes of vocabulary collection, and would either need is own suspension (e.g. TapeIntersect) or a refactoring of the tape/vocab system to make it more modular.

I don't really want to rethink the tape/vocab hierarchy to account for a grammar we don't really use.  It's used in tests, but they were mostly (a) duplicated by Join tests, (b) skipped, or (c) hunting down old semantic issues that are no longer of interest.

There was one remaining test type that's still of interest, the intersection of negations in filters.  That's nontrivial and important (although it happens that the source programmer can't actually do that).  I changed these to joins (filter conditions are single-tape so it's equivalent), and changed the filter-rescoping so it applies to Joins rather than Intersects.