TapeRename bugfix and test updates
I hadn't run the whole test suite in a while, and we had accumulated a bunch of red tests.

(1) Most were just consequences of small changes to the semantics, like what tapes are defined during erroneous matches, or what things count as errors.  I updated the tests to reflect the current semantics.

(2) The rest were due to a bug in how TapeRename handled same-tape renames (e.g. `Rename(X, t1, t1)`).  It was adding the fromTape `t1` to the tape list, and then promptly deleting the toTape, also `t1`.  Obviously these have to happen in the opposite order.