Threading static vocab through ReplaceBlock
This one is a little tricky because the structure of ReplaceBlock is so complex.  I don't love this solution, because we're basically recapitulating the structure we're about to make in a future pass.

Technically that structure is tape-independent, so we could actually make it BEFORE this pass and then just read the tapes off that structure.  But on the other hand, keeping it as a ReplaceBlock until tape inference lets us give some better error messages.

A midway solution could be to make the structure beforehand so we can use it here, but wrap it in a vestigial ReplaceBlock structure that can do error checks at the same time.