Static atomicity
The previous commits didn't take into account atomicity; now I'm taking a stab at it.  This ended up requiring some refactoring of the Vocabs.

(1) The suspension types (VocabSum etc.) ended up being unnecessary; I took those out.  (I may have done this in a previous commit, we haven't needed them in a while.)  If the symbol isn't yet resolved into literals, tapes don't have vocab objects at all, so there's no opportunity to suspend anything.

(2) The old idea of a VocabSet (that's just a set of strings and a wildcard boolean is back) evolved into two different types in a previous commit, VocabAtomic and VocabString.  But it ended up being easier to factor out what they had in common (a set of strings and a wildcard boolean) into their tokens member... meaning that VocabSet is back, just under a different name, WildcardSet.

(3) According to the value of `env.opt.optimizeAtomicity`, calculateTapes returns either VocabAtomic or VocabString.  VocabAtomics undergo fission into strings if any of the following happens:

   (a) They undergo sum/product/intersection with a VocabString
   (b) They've undergone both concatenation and join.
   (c) They overflow beyond some high number of vocabulary items.  (Like if there are 50,000 roots, we don't want a vocab of 50,000 things, we want to break that down into characters.)

This all happens within the vocab system, the tape system doesn't have to know about it.  The only big change to the tape system is that we need a new suspension, TapeProduct, to handle concatenations (in order that (b) can be properly determined).

(4) This mostly works, the only issue so far is getting both the token inventory and wildcard status correct in ReplaceBlock.  I'll leave that to the next commit.