Fixing ReplaceBlock vocabs
(1) The ReplaceBlock vocabs are a bit dicey to get right, and requires a bit of hacky effort rather than simply being read off the structure.

We've got both tapes of Replace having wildcards, because that's how it appears to behave in (for example) testGrammarReplace.  But that's not completely true, the output isn't really "wild", it's determined by whatever the input vocab ends up being.  We don't really have a mechanism for this kind of cross-tape dependency (besides Rename, but that isn't quite right for this).

One way to get it right is just to grab the output vocab from the rule's .toGrammar, which doesn't have a wildcard and thus doesn't improperly propagate wildness.

The resulting structure gets the vocabs and wildness right, but doesn't manage to propagate atomicity properly.  There's actually nothing in this structure that requires VocabString!  I handled this by putting a VocabString WILDCARD in a place that propagates string-ness to everywhere that needs it.

I think perhaps a more principled way of doing that would be to treat the rule as a Repeat -- since it is -- which would give it concat-ness and that (with the join) requires string-ness.  But there's a problem with repeat that I have to get to first.

(2) We're not actually getting the semantics right where join-ness and concat-ness result in string-ness.  I wrote some new tests to show this.

First, we've been assuming "if a tape has been both concatenated and joined, it can't be atomic".  That's not actually true, like yes `Join(t1..., Seq(t1...,t1...))` can't be atomic, but `Seq(t1..., Join(t1...,t1...))` can.  We can express the correct rule but not the way we've been doing it.

Second, our way of calculating it is bugged.  When an intersection would cause the atoms to undergo fission, the way we've written it FIRST calculates the intersection of the vocabs before actually splitting the atoms, resulting (often) in an empty vocab.  But the second issue is a moot point because this code will disappear when we address the first issue.