Using static vocabs in generation
(1) Through the previous changes, we hadn't actually been using the new vocab objects.  We were just testing them, but when it came to actual generation we were still using the old ones.  This commit stops calculating the old ones and just grabs the new ones from the root grammar.

With a few bugfixes (2 & 3), this works for all our source grammars, but there's a deeper issue (4) that we need to solve before moving on.

(2) Vocab calculation for Short wasn't quite right, it needed to be forced into being a VocabString.

(3) Same with Negation, with the additional requirement that it needed to be a wildcard vocab as well.

(4) There's a deeper problem with Replaces (and, probably, Match, but I haven't yet written a test that proves it).  This doesn't actually come up in real sources because Replaces and Matches are always contained in ReplaceBlock, and ReplaceBlock figures out the gross tape/vocab structure on its own.  But it comes up when we test Replaces in isolation, they can't always get their output tapes correct.

The issue is basically the same problem that VocabCopyEdges solves -- that some tapes' vocabulary depends on other tapes, and other tapes' vocabulary has wildcards, which depend on those tapes' total vocabulary at the end of staging.

We've been saying that Replace's output tape ($o) has a wildcard.  But that's not true, not in the way we've been using wildcards.  Replace's output vocab wildcard doesn't mean "Any character in the output tape's vocab".  A replace's output vocab just contains its input vocab... and that vocab contains a wildcard.  But that's a wildcard reference to the input tape's vocab, not to the output tape's vocab!

We're copying over that wildcard but what it actually means changed.  Basically, wildcards have to be relativized to a tape.  That's what they are -- they're vocab references.

VocabCopyEdges are basically the same idea underneath.  When that algorithm said "I found a copy edge between t1 and t2", what that means in our new parlance is "t2's vocab contains a wildcard reference to t1's vocab."  I don't think we want to use VocabCopyEdges, however.  I'm pretty sure it would be subject to the same Big Semantic Issue we're trying to solve in these branches, and I could construct a failing test using similar ideas to our failing Cursor tests.

Probably we need something like the reference/suspension/unification system used in Tapes.  I had hoped that we could avoid having BOTH tapes and vocabs use a reference/suspension/unification system, and only use Tape's system, but I don't think that's possible.  We can't fully know the vocabs before staging, and we can't suspend tapecalc that long, we need to nail down the tapes for certain much earler than that.