Re-doing wildcard system
We determined that the previous wildcard system (in which there was just a nonspecific wildcard flag meaning "And it might have other vocab") was flawed, because it didn't manage to capture what's going on in Matches (and thus Replaces).

This is a beginning to that, although there are still some bugs to work out.

(1) Wildcards are gone; in their place are `VocabRef(tape)`, meaning "Any character in the vocab of `tape`".

(2) The Vocab system now has suspensions and unification just like the Tape system.  Note that handling renames is a bit tricky, because they change the "symbol" table the vocabs use in order to resolve references to Tapes.  I'm not sure whether the way I'm doing it is the best way.

Suspensions are built immediately, but they're not resolved until the last minute, because querying may change the answer in the end.

(3) We often weren't catching thrown exceptions from Interpreter creation in the test suite, making it hard to find which tests were throwing them.  We now catch them in both prepareInterpreter and sheetFromFile.

(4) Using magic strings for tags (e.g. "VocabAtomic") had the downside that VSCode's automatic symbol renaming can't replace those throughout the code.  If you ever want to change it, you have to find-n-replace.

In this commit, I'm experimenting with using a `const enum` instead, like `VTag.Atomic`.  While enums are objects and require lookup, const enums are inlined.