Finishing static vocabs
With some bugfixes and a lot of refactoring, this puts us back to green except for the Cursor tests we've been working towards.

(1) A few commits ago I said that we didn't need a special kind of suspension for Matches, we could just use renames.  This isn't true.  While this gets the simpler matches right, it doesn't get the difficult case right, and the difficult case is the one that matters.

   Join(t1:hello, Match(t1:.*, t1, t2))

If we say that Match's vocab contains a TapeRename of the child's tape set from t1 to t2, that creates a VocabRename that changes what the t1 wildcard inside refers to: it now refers to the vocabulary of t2, just like it would in `Rename(t1:.*, t1,t2)`.  This is the right interpretation for Renames, but not for Matches -- they absolutely do need to refer to t1, that's the whole point of cross-tape wildcards.

So really, Matches do need their own suspension.  In the end, what they do is much simpler than Renames, because they don't change the interpretation of anything, they just copy the vocab.

This is also what Replaces use.

(2) Moved to a const enum for Tape tags to avoid magic strings.

(3a) It was getting a little cumbersome to type out a bunch of TapeX and VocabX.  But then I realized, this was basically me manually namespacing, to keep the parallel TapeX and VocabX pairs from clashing.

JS/TS already does this for us; modules are namespaces, so even if we name everything X, we can `import * as Tapes from...` and it'll be Tapes.X automatically.  (And when we want something directly or use it all the time, we can still import it like `import { X, Y, Z } from...` even if you've done the above.  I don't want to call Tape a `Tapes.Tape`, so I just import it directly.)

This makes some things read cleaner and others messier, but I feel like some of the places where it feels messy is actually due to bad encapsulation and naming on my part.  Tape and Vocab should ideally be a bit more self-contained so that fewer things should have to know about their internal business.

(It would be good, for example, to be able to have things actually be typed Lit when they're resolved, rather than continuing to have to check their tags all over the place.  But every way I can think of doing this results in extreme Typescript jank.)

(3b) Way too many things are named just "tape" or "tapes".  I've tried to address this a bit.

The thing inside a TapeLit used to be called "tapes", because it was just a set of strings with the tape names.  But now it's a mapping from tapeNames to Vocabs, so it's vocabMap now.  Same thing inside vocab unification, that's a vocabMap.

We barely use the old Tape object and it's about to disappear, so for now it's OldTape.  Then I renamed TapeInfo to Tape -- that's a better use of the plain word "Tape" because it's the basic tape object now.

The tape getter `Grammar.tapes` sounds ambiguous, do we want the actual tape object or just the names?  It's actually the latter, so this is now `Grammar.tapeNames`.

Cursor.tape is now Cursor.tapeName to align better with Literal, Hide, etc.

(4a) The tags in vocab were expressing two different things -- one, was it a suspension or something concrete, and if it was concrete, was it atomic?  This made testing for literalness cumbersome, unlike Tapes.  (The way tapes did it was inspired by a previous Vocab system, but Vocab changed since then.)

Now these are separate.  `.tag` is about being a lit/ref/suspension, and `.atomicity` is the atomicity class.

(4d) VocabSeq and VocabString were always a bit misleading even more misleading when they became Seq and String.  That's not a Seq, that's not a String!  They're now "Concatenated" and "Tokenized".

(5) Some time back, I renamed `resolve` in Tapes and Vocabs to `unify`, but I'm not sure this really counts as a unification algorithm, so I'm changing it back.