Refactoring and renaming
(1) In the last commit I changed the basic tape object, the one every Grammar has, to be called Tape.  But this is actually misleading, it contains multiple tapes.  So it's now a TapeSet.

(2) I think I forgot to mention this in the last commit message, but Tapes.resolve and Vocab.resolve were both post-order traversals and a lot of the resolution code was just traversal boilerplate -- perform this operation on my children and then update myself.  I made a Pass-like traverser that uses a map() function instead.

I didn't actually make them Passes, though, there's still a lot in Passes that's irrelevant to them.

(3) This got me thinking about what functionality *shouldn't* be in Passes.  Fundamentally, they need to express Recursion Strategies.  Error message handling probably needs to be in there too, that's one of the reasons we use recursion strategies in the first place, is how tedious and error-prone it is to handle this manually through a recursive traversal.

On the other hand, the ability to time themselves and log to the console has nothing to do with traversal.  We only do it at the root.  In fact, when we DON'T want the ability to time and log we ended up calling `Pass.transform()` directly. (But sometimes we forgot and still called `Pass.go()`, because who would think that timing/logging was the difference between "going" and "transforming"!?!)

This gets rid of `Pass.go()`, all there is is `.transform()` now, and introduces a TimerPass that takes a child Pass, times it, and logs it.

(4) This also got me thinking that "Result" is a terrible name for a messaging monad.  Who would think that's what "Result" means?  And we use the word "result" in lots of places that aren't messaging-related, like... results of things.

So Result<X> is now Msg<X>.  (Msg in turn is now Message, and Msgs is gone, it's now just Message[].  We don't actually use this word that often in the code anymore, not that we have so much automation of messaging, so saving those few characters doesn't really matter.  But we still have Results everywhere, so that gets the abbreviated form.)

(5) Last week (or so), I gave Passes the improved message handling of the former CatchingPass, letting us throw a Msg<X> rather than return it (and thus simplifying a lot of return signatures).  But I didn't actually change any Passes to use it, so while I was renaming Msgs I adjusted some error handling to take advantage of it.