Threading Expr.forward through Concats
Property-testing the new tape/vocab system revealed that we had never threaded `Expr.forward` through ConcatExpr.

The test case is a bit strange, but simple: A Seq of two embeds, both referring to epsilons.  Nothing embedded is automatically simplified, you have to visit it with a forward, delta, or deriv.  But deriv and delta are never called because there are no tapes in this grammar, and forward can't get through the Concat, so nothing can simplify those embeds, so the whole thing can't be simplified.

The unnerving questions is why this never failed BEFORE?  It's a rather fundamental failure, nothing has changed about property testing since the last time I've run it, and I haven't touched Exprs in months.

The only thing I can think of is that we had an error in tape calculation before, and in the process of fixing tapecalc I've squashed that bug.  If we were inferring the existence of a spurious tape, the resulting delta on that tape would suffice to simplify the tree.  (In fact, in the pre-forward days, we used to do that as a little hack to achieve just that.)

Anyway, I fixed this up.  This also suggests that for any kind of Expr that doesn't implement forward, the property test framework will find a failing test with the same problem.  But I'll get to that later, I want to shore up tapes first.