Starting to use local tape/vocab instead of global
(1) DerivEnv has vocab/atomic members now, and so do Cursor, with the idea that Cursor will be the one providing those to the env.

The idea is that we'll use this rather than a TapeNamespace, which
    (a) Doesn't allow for non-global tapes, which is a near-term goal.
    (b) Is probably part of the inefficiency of deeply-nested rewrite rules.

(2) But before making the Cursors actually do this heavy lifting, we have to make some decisions about what exactly to do about Cursors that reference tapes that don't exist.  Should they disappear?  With an error warning or without?  What about Cursors with no vocab?

I could have just made them quietly not exist and that would probably have worked fine, but that goes against our current goal of no-longer-being-loosey-goosey-about-what-tapes-exist-where.  So I'm going to spend some more time investigating each missing tape error and make sure there's not something of concern.

(3) One such situation is when there's a Correspond but its input or output are Epsilon, and then there's a manually-made Cursor around it.  Corresponds don't actually put their input/output tapes into tapecalc, so that Cursor ends up referencing a non-existent tape.

Investigating this, though, I found a bug.  It's not something that can arise in real source grammars right now, because CorrespondGrammars are currently only used in testing.  But it can cause infinite generation so we have to squash it.

If we DON'T manually put a Cursor around such a Correspond, it doesn't get one during CreateCursors (because the relevant tape doesn't exist), but without actually doing deriv/delta on the tape, Correspond doesn't actually finish, so we can get stuck in infinite generation.

So either Corresponds have to disappear when tapecalc finds one of their tapes doesn't exist, or they have to ensure their tapes exist.  I opted for the latter.  Correspond should probably act like Short and Not anyway, it should force tokenized interpretation on its tapes' vocab.  The way we do that makes the tape exist, so that makes our decision for us.

(4) Removed `Grammar.getChildren()`, it's just a wrapper around `getChildren(g)`.  (Then for brevity just renamed that `children`.)