Attempting to get vocabs locally resolved
I'm trying to make it so that Cursors can simply look at their own `.tapes` object and read the vocab they need.  In the last commit, I was on the track of some minor bugs where they refer to tapes that don't locally exist.

Here, I'm looking at a much trickier situation, because it involves a bit of a paradox.  I think I know the solution but want to commit some code and tests illustrating the problem first.

(1) For a Cursor to grab a local vocab, it has to be resolved.  Before this, we've only resolved vocabs at the top level, so if the Cursor already exists in the tree (as many explicit-priority tests do!), its local vocab won't be resolved.

So now, we thread resolution through the tree itself.

(2) I was hoping we might be able to *not* resolve every vocab in the tree (because there are lots), but only resolve the vocabs exactly at Cursors.

(3) Here, we run into a bit of a snag.  Consider  `Cursor(t1, Cursor(t2, X))` where X contains a situation like a Match resulting in t1's vocab having `Vocab.Ref(t2)`.

Okay, now we try to resolve t1 at its respective Cursor.  It contains a ref to t2... but t2 shouldn't exist here, tapes don't exist outside their cursor.  (I mean, they do AT THE MOMENT but they won't soon.)

This suggests we have to bite the bullet and do one last tapecalc to resolve *every* vocab in the tree.  The resolved child vocab of that inner cursor would no longer have `t1: Vocab.Ref(t2)`, it would just be those strings.