Letting Envs vary by the pass
There's an abstraction about components, passes, and envs that we weren't capturing, and in fact were doing something kind of silly.

This was making it tedious to adapt CalculateTapes in the way that solves the last commit's issues, so I figured it was high time to fix this.

(0) Before doing this I switched back to using top-level vocabs in generation, to return us to (mostly) green, so I could refactor in safety.

(1) Out of suspicion, I deleted `CollectionGrammar.mapChildren` (which is responsible for putting its symbols into env.symbolNS), which should have caused major failures left and right.  That's part of WHY mapChildren is a member, so that CollectionGrammar can put those there, because that's really important, right?  That's part of why we have a symbolNS in PassEnv, right?

Turns out, no.  Nothing failed because we never BOTH used mapChildren AND used those symbols from it.  All the remaining functions that use `env.symbolNS` that way aren't component-to-component Passes, so they can't use mapChildren anyway.  They were changing env.symbolNS and then recursing manually.

Meanwhile, there are places where we're manually changing parts of the environment before recursion, but not using Env, because Env is quite limited in what it can do.  If we need to do anything in a Pass that isn't push symbols onto the symbolNS, we can't use the boilerplate and thus have to manually recurse -- and as mentioned, pushing symbols isn't even something we need during the main passes.

So really, Env wasn't doing much.  It was really just a place to store Options.

(2) Stepping back, what we need Env to do is to assist in the tedious management of certain kinds of information during a recursion strategy: stuff that mostly stays intact through the traversal but occasionally changes.  Symbols, tape references, etc.

The reason Env wasn't very good at this is that different passes have different environmental needs -- they need different information, and change it in different ways.  But the way we create and compose them made it difficult for them to vary.  We start with a PassEnv at the very beginning and then thread it through ~20 passes.

(3) Do passes ever need to share an Env?

Sometimes!  ExecuteTests is a good example of this.  Now that staging is handled by Passes, and ExecuteTests does its own staging, that means ExecuteTests now has 4 or 5 "child" passes.  Those child passes are executed on nodes that aren't themselves roots, so they won't necessarily go through a CollectionGrammar to get something into a symbol table.  Therefore yes, they need to use their parent's env, which DID go through a CollectionGrammar.

On the other hand, we never have two "sibling" passes -- ones that happen one after another -- that need to share information.  We've been cautious to avoid that, we try to share that kind of information only by adding something to the tree.

So `Pass.compose` doesn't have to thread an env through both of its children.  NOT threading the env like this takes away the need for them to all be the same type.

(4) While Env-changing takes various forms, it's typically done in the same place -- immediately before recursing.  (I mean, if we changed it *after* recursing nothing would happen.  It's not like we return them, if we made a new env it would just be garbage collected when we leave the scope.)

So now there's a method `Env.update(g)`, basically meaning "Change the env in whatever way necessary using information from g".  By default it doesn't change anything; if you need to change something, inherit from Env and override this method.

For example, `SymbolEnv.update(g)` looks at g, and if it's a CollectionGrammar, it pushes those symbols into the `SymbolEnv.symbolNS` and returns the new env.  Otherwise it returns the original env.  This will be called automatically by mapChildren -- achieving what `CollectionGrammar.mapChildren` used to do, but now only when necessary, and when something else is necessary, `WhateverEnv.update(g)` can do that instead.

(5) Passes now have a method `.getEnv(opt)` that'll provide an empty instance of whatever kind of Env that pass needs, and the requests opts.

(6) On Friday-ish I got rid of `Pass.go()` for being both useless and poorly-named.  But we now need a top-level method in the same place.  We don't want to use `.transform(g, env)` everywhere because that's passing in an env, and except in special circumstances we AREN'T passing in envs, we're getting new ones with `.getEnv(opt)`.

So now there's a method `.getEnvAndTransform(g, opt)` that does what it says on the tin.  This is the method that CombinedPass calls instead of calling transform, and the main one we'll want to be calling elsewhere.

(7) I think we can simplify all this more in the future, but anything more than this is premature refactoring, we have to make sure this does what CalculateTapes is going to need, first.