Making tapecalc use new Env system
This is a pretty simple change but I think the general idea will bear fruit with vocabs later.

Before, Embeds looked in the `.tapeMap` for their referents.  This was initially empty, so if they couldn't find one they made a placeholder.  Now Env and CollectionGrammar make those together, filling a table full of tape refs.  Embed assumes it'll find one in there, and grabs it.

For tapecalc this isn't a big deal -- there's only one place it happens -- but I think it'll pay off for vocabcalc, because vocab refs are generated in more places.