Bug in replace otherContext
A previous commit broke one of our replace tests.  It's something to do with otherContext.

New tests in testPassTapes reveal that Replace and ReplaceBlock report different vocabularies for otherContexts -- Replace thinks otherContexts must be tokenized, ReplaceBlock that they can be atomic.  And the ReplaceBlock -> Replace transformation happens in between the vocab being calculated and being used.  That suggests that this is the culprit.

I also made a change to testGrammar, allowing the programmer to specify vocabs either just by number like `{t1: 3}` or explicitly like `{t1: ["h","e","l"]}`.  That way when we want to be REALLY sure what a vocab contains, we can check.