Fixing Replace/ReplaceBlock tape discrepancy
This wasn't the culprit of the previous bug, but it's still good to fix while we're at it.

Back when I hadn't quite figured out the semantics of wildcards, I had gotten the vocabulary of Replace subtly wrong.  Because of this, definiting ReplaceBlock's vocab compositionally from its component rules didn't work, and I figured out a complex workaround.  But this also meant they could get different answers in other ways, like in how they handle otherContext.

Now the vocabulary of Replace has been corrected, letting us get rid of the complex workaround in ReplaceBlock and use the actual vocab of Replace.  This makes it so they have the same opinion about otherContext.