Sorting/stringifying test outputs before comparing
In property tests, we sort and stringify StringDicts, and then sort the resulting arrays, before comparison.  I ported this back to testMatchOutputs.

This makes it considerably easier to read the "+expected/-actual" diffs.  The diffs are order-sensitive even if the comparison wasn't, and give that StringDicts print out over multiple lines, it meant that unrelated entries could be "paired up" for comparison.  This puts each StringDict on its own line (because for comparison it's a string), and makes it so you don't get a spurious +/- pair for content that's in a different order than expected.

It also seems to speed up comparison considerably, NotMatch tests now execute in ~50ms.