Fixing otherContext bug
(1) The answer was simple in the end.  Yesterday I had stopped properly threading the Options through `.tapify(env)`, and so the grammar and the query were tokenizing things in different ways.

(2) Also, even when the query was being tokenized as Tokenized, it was using a Vocab.Atomic object to do so.  (This wasn't the cause of this particular bug, though.)