Getting vocabs resolved at cursor level
This is an attempt to make sure that Cursors have the correct vocabs by the end, by having them resolve their cursors during tapecalc itself.

I don't love this solution, because at the end we have to redo tapecalc entirely, which I feel goes against some of the goal of static tapes.  But if we separated out all vocab-related calculation into its own late pass, it'd require calculating a lot of the same tape information anyway, so I think it'd look very similar in the end.

I think we have to accept for now that tapecalc and vocabcalc are closely intertwined, and this structure can only FULLY be resolved at the end... but meanwhile we need some preliminary tape information from it earlier, and so end up running it before it's fully "ready".

This doesn't yet USE these resolved vocabs anywhere, it's still using the vestigial collectAllVocab/TapeNamespace solution; this is just to make sure they're in place for the next step.