Using cursor vocabs instead of global
(1a) CursorExprs now push their vocab info into env, and exprs that need it like DotExpr (which needs the vocab) and LiteralExpr (which needs atomicity) use that instead.

(1b) This largely works, but it does NOT allow input/output vocabs to differ in matches or replaces (because in the absence of the global TapeNamepsace there's no way for MatchExpr to grab another tape's vocab).  We've decided that these shouldn't differ anyway, but this changes the answers of MANY match/replace tests.

(2) Our tests varied whether they used `Any("t1")` or `Dot("t1")`.  The parameterization of `Any` is a subset of that of `Dot`, so they've all been changed to `Dot`.

(3) One consequence of the new Cursor vocab system that we'll have to remember in testing: You can't query on a tape where there's an explicit Cursor.  Queries are outside of that, and tapes shouldn't exist outside of their cursors.