Removing TapeNamespace system
(1) OldTape, TapeNamespace, collectAllVocab, and everything that used/referenced them had become vestigial; they no longer made any difference to the actual results.  So this commit eliminates them.

This changes the parameterization of many things -- basically, we were keeping vocabs globally and having to thread them through to everything that had anything to do with vocabs, and we can now dispense with that.

(2) The ResolveVocab pass wasn't doing anything but making some redundant checks (which will be made in a moment by constructExpr anyway), so that's gone too.

(3) This throws a big wrench into vocab testing.

The problem is that the conceptual object we want to test no longer exists -- the vocab that is both resolved AND exists at the top level.  Vocabs are now guaranteed to be resolved only by their respective Cursors, and once we add that cursor, the vocab doesn't exist outside of that Cursor.

Vocabs without any Refs (that is, "wildcards") happen to be resolved earlier than the addition of their Cursor, and so all of those tests still pass, but a lot of the more interesting/difficult vocab tests don't.

There's not much we can do about this -- our tests, by their nature, test information that exists at the Interpreter/root and will have trouble testing any information that exists only within the tree.  That's a long-term concern as we're slowly moving information out of the top-level/root into the tree, because those things become opaque to testing.  In the long run we may want to have special AssertionGrammars that we can stick into the tree to assert things about their local structure.