Cursors now tape-opaque
BACKGROUND: We had the question long ago, what does it mean if you say `Cursor(t1, X)` and t1 already exists outside of this scope?

  (a) Forbid this.  We didn't want to, however, because ultimately we want something along these lines for good encapsulation -- hidden tapes, pretapes, etc. -- without necessarily going through and renaming everything to unique names.  Unique names are a bit tricky, too, in recursive situations -- even if you rename it, you can get into this situation recursively.  It's better to say "This is a possible configuration, now what does it MEAN?"

  (b) It means both t1s are the same tape.  But if one tape can be associated with two Cursors (as would be necessary to handle material outside this Cursor), you end up with nonsensically scrambled results.

  (c) The t1s just refer to different tapes.

We decided on (c) and made generation work like this, but there was a problem: basically everything else assumed (b)!  Tape priority assumed (b), vocab collection assumed (b), etc.  This meant that two different Cursors could reference two supposedly different tapes, yet still have the vocab of one affect the vocab of the other, undermining encapsulation.  You could make it so joins intersected to the wrong tapes, have un-Cursored material, etc.

Making this true required a lot of re-doing of how tapes and vocabs work, because tapes and vocabs were things that existed globally (or to be more precise, belonging to a member of Interpreter), which basically requires (b).  Making (c) true across the board means giving up on that and having tapes/vocabs be things that exist only in the tree.

But a lot of this work was stuff we wanted to make anyway.  We had long been loosey-goosey about "What is the tape/vocab of this at this point in time?", recalculating tapes from scratch dozens of times, in part because we only had sloppy ideas about what the semantics of tapes were.  Now we know somewhat better and want to nail it down as early as possible.

(1) CursorGrammar.tapes no longer contains its own special tape.  This is ultimately necessary to get things like Joins correct; the parents in the tree are looking at their childrens' `.tapes` to make decisions like this.

(2) To make this work requires another tape suspension that removes a tape from a TapeSet.

(3) This, of course, changes what tapes exist in various tests, so I fixed those up.

(4) The vocab resolution system as it existed didn't suffice to get everything correct.  No matter what I put into the vocab map or when (pre or post-traversal) I performed the resolution, I could construct some test where it failed.  (Why didn't it fail before?  Well, because tapes and their vocab still existed outside of their cursors, which gave us some wiggle room.)

Going back to an older resolution algorithm worked in the end, although II don't think we ever used in on vocabs.  (What I once called a "push" algorithm, eagerly iterating through all the vocabs and replacing every instance of a key (say, "t1") with its referent, rather than a "pull" algorithm where the vocabs resolve themselves by looking up the keys they contain from `env.vocabMap`.)  And now that it works, it makes sense why "push" works.

    (a) The "pull" algorithm required us to construct a vocabMap that contained all the information necessary to resolve any vocab.  We had that when tapes/vocabs were global.  Once references to tapes could only exist inside their Cursor's scope, however, building a vocabMap that contains the right information everywhere and at every time it's needed, without violating cursor opacity, was simply too difficult.

    (b) Really, what we needed was the ability to go through a vocabMap and make sure to resolve any references that were about to become "illegal".  And that's the "push" algorithm in a nutshell.

Cursor is the clear locus of this responsibility, like however it happens, Cursor_t1's child's tapes can contain vocab references to t1, but Cursor's tapes cannot, because Cursor's parent will look at it and say "Huh?  There's no such tape!"  And cursor has enough information to make that happen, it's the outermost place where t1 occurs, so nothing adds anything to t1 except in its scope.

I had been concerned back then that "push" would be inefficient because it was operating on the product of symbols and tapes, but that's when I was envisioning this happening at a top(ish) level, like while tapecalcing CollectionGrammar.  Now that tapes are local it doesn't come up in that way.

Since CursorGrammar.tapes doesn't contain its special tape anymore, we need somewhere to store the resolved vocab object so that it can be passed to the CursorExpr.  So now there's a member CursorGrammar.vocab that contains exactly this.

(5) TODO: There's no code in the new push resolution to handle renames correctly, fix that.

(6) As I suspected last time, many match/replace tests did indeed break again, once the tape/vocab semantics was fixed.  We'll have to go through those to decide which are still testing things we care about.